feat: rename generated file to have bg instead of wasm
