+++ examples/hello_world/chrome/index.js
@@ -6,6 +6,6 @@
 // wasm (generated by wasm2es6js) which has a `booted` promise to let us know
 // when it's ready to go.
 import { greet } from './hello_world';
-import { booted } from './hello_world_wasm';
+import { booted } from './hello_world_bg';
 
 booted.then(() => greet("World!"));
