+++ examples/smorgasboard/.gitignore
@@ -1,3 +1,3 @@
 package-lock.json
 smorgasboard.js
-smorgasboard_wasm.wasm
+smorgasboard_bg.wasm
